requirements.txt
langgraph 
langchain-google-genai 
e2b-code-interpreter 
pydantic
python-dotenv
IPython